+++ sjsonnet/src-jvm-native/sjsonnet/CachedResolvedFile.scala
@@ -4,7 +4,6 @@ import java.io.{BufferedInputStream, File, FileInputStream}
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 
-import net.jpountz.xxhash.{StreamingXXHash64, XXHashFactory, XXHash64}
 import fastparse.ParserInput
 
 /**
@@ -65,34 +64,10 @@ class CachedResolvedFile(val resolvedImportPath: OsPath, memoryLimitBytes: Long,
   override lazy val contentHash: String = {
     if (resolvedImportContent == null) {
       // If the file is too large, then we will just read it from disk
-      CachedResolvedFile.xxHashFile(jFile).toString
+      Platform.hashFile(jFile)
     } else {
       resolvedImportContent.contentHash
     }
   }
 }
 
-object CachedResolvedFile {
-  val xxHashFactory = XXHashFactory.fastestInstance()
-
-  def xxHashFile(file: File): Long = {
-    val buffer = new Array[Byte](8192)
-    val hash: StreamingXXHash64 = CachedResolvedFile.xxHashFactory.newStreamingHash64(0)
-
-    val fis = new FileInputStream(file)
-    val bis = new BufferedInputStream(fis)
-
-    try {
-      var bytesRead = bis.read(buffer)
-      while (bytesRead != -1) {
-        hash.update(buffer, 0, bytesRead)
-        bytesRead = bis.read(buffer)
-      }
-    } finally {
-      bis.close()
-      fis.close()
-    }
-
-    hash.getValue()
-  }
-}
